cleanup dead codes
